Changes required to be batch-compatible
The only changes required for the proof to be batch-compatible are
that the proof, instead of consisting of `Gamma || c_string || s_string`,
it now consists of `Gamma || U_point || V_point || s_string`.

The tests vectors are changed, but only to reflect the above. It can
be verified that the values of `U_point` and `V_point` are taken
from the test vectors of the draft (where they provide also these
points).